WIP: Turned adjusting contact mechanics parameters into a modal window
